migrating context to redux
