+++ src/components/sign-in/sign-in-form.component.jsx
@@ -6,7 +6,7 @@ import {
 } from "../../utils/firebase.utils";
 import FormInput from "../form-input/form-input.component";
 import Button, { BUTTON_TYPE_CLASSES } from "../button/button.component";
-import { UserContext } from "../../context/user.context";
+
 import {
   SignInContainer,
   SignInContent,
