--- src/context/categories.context.jsx
@@ -1,34 +1,0 @@
-/** @format */
-/* eslint-disable react-refresh/only-export-components */
-
-import { createContext, useState, useEffect } from "react";
-import { getCategoriesAndDocuments } from "../utils/firebase.utils.js";
-import SHOP_DATA from "../shop-data.js";
-
-export const CategoriesContext = createContext({
-  categoriesMap: {},
-});
-
-export const CategoriesProvider = ({ children }) => {
-  const [categoriesMap, setCategoriesMap] = useState({});
-
-  // useEffect(() => {
-  //   addCollectionAndDocuments("categories", SHOP_DATA);
-  // }, []); // run only once on mount
-
-  useEffect(() => {
-    const getCategoriesMap = async () => {
-      const categoryMap = await getCategoriesAndDocuments();
-      setCategoriesMap(categoryMap);
-    };
-    getCategoriesMap();
-  }, []);
-
-  const value = { categoriesMap };
-
-  return (
-    <CategoriesContext.Provider value={value}>
-      {children}
-    </CategoriesContext.Provider>
-  );
-};
