+++ src/main.jsx
@@ -6,21 +6,16 @@ import { BrowserRouter } from "react-router-dom";
 import { Provider } from "react-redux";
 import "./index.scss";
 import App from "./App.jsx";
-import { UserProvider } from "./context/user.context";
-import { CategoriesProvider } from "./context/categories.context.jsx";
+
 import { CartProvider } from "./context/cart.context.jsx";
 import { store } from "./store/store";
 createRoot(document.getElementById("root")).render(
   <StrictMode>
     <Provider store={store}>
       <BrowserRouter>
-        {/* <UserProvider> */}
-        <CategoriesProvider>
-          <CartProvider>
-            <App />
-          </CartProvider>
-        </CategoriesProvider>
-        {/* </UserProvider> */}
+        <CartProvider>
+          <App />
+        </CartProvider>
       </BrowserRouter>
     </Provider>
   </StrictMode>
