+++ src/routes/categories-preview/categories-preview.component.jsx
@@ -1,11 +1,12 @@
 /** @format */
 
-import { Fragment, useContext } from "react";
-import { CategoriesContext } from "../../context/categories.context";
+import { Fragment } from "react";
+import { useSelector } from "react-redux";
+import { selectCategoriesMap } from "../../store/categories/category.selector";
 import CategoryPreview from "../../components/category-preview/category-preview.component";
 
 const CategoriesPreview = () => {
-  const { categoriesMap } = useContext(CategoriesContext);
+  const categoriesMap = useSelector(selectCategoriesMap);
   return (
     <Fragment>
       {Object.keys(categoriesMap).map((title) => {
