+++ src/routes/category/category.component.jsx
@@ -1,14 +1,15 @@
 /** @format */
 
 import { useParams } from "react-router-dom";
-import { useContext, useEffect, useState, Fragment } from "react";
-import { CategoriesContext } from "../../context/categories.context";
+import { useEffect, useState, Fragment } from "react";
+import { useSelector } from "react-redux";
+import { selectCategoriesMap } from "../../store/categories/category.selector";
 import ProductCard from "../../components/product-card/product-card.component";
 import { CategoryContainer, CategoryTitle } from "./category.styles";
 
 const Category = () => {
   const { category } = useParams();
-  const { categoriesMap } = useContext(CategoriesContext);
+  const categoriesMap = useSelector(selectCategoriesMap);
   const [products, setProducts] = useState(categoriesMap[category]);
   useEffect(() => {
     setProducts(categoriesMap[category]);
