+++ src/store/categories/category.selector.js
@@ -1,0 +1,3 @@
+/** @format */
+
+export const selectCategoriesMap = (state) => state.categories.categoriesMap;
