+++ src/store/categories/category.types.js
@@ -1,0 +1,5 @@
+/** @format */
+
+export const CATEGORIES_ACTION_TYPES = {
+  SET_CATEGORIES_MAP: "category/SET_CATEGORIES_MAP",
+};
