+++ src/store/user/user.reducer.js
@@ -1,5 +1,5 @@
 /** @format */
-import { USER_ACTION_TYPES } from "./user.type";
+import { USER_ACTION_TYPES } from "./user.types";
 
 const INITIAL_STATE = {
   currentUser: null,
