+++ src/store/user/user.types.js
@@ -1,5 +1,5 @@
 /** @format */
 
 export const USER_ACTION_TYPES = {
-  SET_CURRENT_USER: "SET_CURRENT_USER",
+  SET_CURRENT_USER: "user/SET_CURRENT_USER",
 };
